Version 1.0 : Explorer App
This version has the below functionality.
On hitting the allow current location access, the app gets the current latitude and longitude.
You can then select the distance radius and hit generate location.
You get 15 seconds to sets your intent. After this the true random number generator generates the coordinates of a location that it within the specified distance radius.